Moved screen names to new file
